requirements.txt
accelerate==1.8.1
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
av==14.4.0
blessed==1.21.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
decord==0.6.0
distro==1.9.0
e==1.4.5
einops==0.8.1
et_xmlfile==2.0.0
fastapi==0.115.14
ffmpy==0.6.0
filelock==3.18.0
flash_attn==2.8.0.post2
fonttools==4.58.4
fsspec==2025.5.1
gpustat==1.1.1
gradio==5.35.0
gradio_client==1.10.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==1.93.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pandas==2.3.0
peft==0.15.2
pillow==11.2.1
portalocker==3.2.0
protobuf==6.31.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.1
safehttpx==0.1.6
safetensors==0.5.3
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sty==1.0.6
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.9.0
timeout-decorator==0.5.0
timm==1.0.16
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.51.3
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
wcwidth==0.2.13
websockets==15.0.1
xlsxwriter==3.2.5
